delete air-level mask checking
