+++ source/vir/src/ast_to_sst_func.rs
@@ -12,8 +12,7 @@ use crate::ast_util::unit_typ;
 use crate::ast_visitor;
 use crate::context::{Ctx, FunctionCtx};
 use crate::def::{unique_local, Spanned};
-use crate::inv_masks::MaskSetE;
-use crate::inv_masks2::MaskSet;
+use crate::inv_masks::MaskSet;
 use crate::messages::{error, Message};
 use crate::sst::{BndX, Exp, ExpX, Exps, LocalDeclKind, Par, ParPurpose, ParX, Pars, Stm, StmX};
 use crate::sst::{
@@ -272,7 +271,6 @@ fn func_body_to_sst(
                 local_decls: Arc::new(termination_decls),
                 statics: Arc::new(vec![]),
                 reqs: Arc::new(vec![]),
-                mask_set: Arc::new(MaskSetE::empty()),
                 unwind: UnwindSst::NoUnwind,
             };
             Some(termination_check)
@@ -567,7 +565,6 @@ pub fn func_def_to_sst(
         MaskSpec::InvariantOpens(exprs) | MaskSpec::InvariantOpensExcept(exprs) => exprs.clone(),
     };
     let mut inv_spec_exps = vec![];
-    let mut inv_spec2_exps = vec![];
     for e in inv_spec_exprs.iter() {
         let e_with_req_ens_params = map_expr_rename_vars(e, &req_ens_e_rename)?;
         let exp = if ctx.checking_spec_preconditions() {
@@ -581,18 +578,13 @@ pub fn func_def_to_sst(
         };
 
         let exp = state.finalize_exp(ctx, &exp)?;
-        inv_spec_exps.push(crate::inv_masks::MaskSingleton { expr: exp.clone(), span: e.span.clone() });
-        inv_spec2_exps.push(exp.clone());
+        inv_spec_exps.push(exp.clone());
     }
     let mask_set = match &mask_spec {
-        MaskSpec::InvariantOpens(_exprs) => MaskSetE::from_list(inv_spec_exps.clone()),
-        MaskSpec::InvariantOpensExcept(_exprs) => MaskSetE::from_list_complement(inv_spec_exps.clone()),
+        MaskSpec::InvariantOpens(_exprs) => MaskSet::from_list(ctx, &inv_spec_exps, &function.span),
+        MaskSpec::InvariantOpensExcept(_exprs) => MaskSet::from_list_complement(ctx, &inv_spec_exps, &function.span),
     };
-    let mask_set2 = match &mask_spec {
-        MaskSpec::InvariantOpens(_exprs) => MaskSet::from_list(ctx, &inv_spec2_exps, &function.span),
-        MaskSpec::InvariantOpensExcept(_exprs) => MaskSet::from_list_complement(ctx, &inv_spec2_exps, &function.span),
-    };
-    state.mask = Some(mask_set2);
+    state.mask = Some(mask_set);
 
     let unwind = match req_ens_function.x.unwind_spec_or_default() {
         UnwindSpec::MayUnwind => UnwindSst::MayUnwind,
@@ -704,7 +696,6 @@ pub fn func_def_to_sst(
             ens_spec_precondition_stms: Arc::new(ens_spec_precondition_stms),
             kind: PostConditionKind::Ensures,
         }),
-        mask_set: Arc::new(mask_set),
         unwind,
         body: stm,
         local_decls: Arc::new(local_decls),
