+++ source/vir/src/expand_errors.rs
@@ -136,7 +136,7 @@ fn get_fuel_at_id(stm: &Stm, a_id: &AssertId, fuels: &mut HashMap<Fun, u32>) -> 
             }
             return false;
         }
-        StmX::OpenInvariant(_, stm) => {
+        StmX::OpenInvariant(stm) => {
             if get_fuel_at_id(stm, a_id, fuels) {
                 return true;
             }
