+++ source/vir/src/sst.rs
@@ -196,7 +196,7 @@ pub enum StmX {
         typ_inv_vars: Arc<Vec<(UniqueIdent, Typ)>>,
         modified_vars: Arc<Vec<UniqueIdent>>,
     },
-    OpenInvariant(Exp, Stm),
+    OpenInvariant(Stm),
     ClosureInner {
         body: Stm,
         typ_inv_vars: Arc<Vec<(UniqueIdent, Typ)>>,
@@ -278,7 +278,6 @@ pub struct FuncDeclSst {
 pub struct FuncCheckSst {
     pub reqs: Exps,
     pub post_condition: Arc<PostConditionSst>,
-    pub mask_set: Arc<crate::inv_masks::MaskSetE<Exp>>,
     pub unwind: UnwindSst,
     pub body: Stm,
     pub local_decls: Arc<Vec<LocalDecl>>,
